Enables use with Homebrew on Linux by breaking some pyenv rules
I'm not convinced of its reliability because of how Homebrew's
pkg-config produces versioned URLs to libraries' CFLAGS and LDFLAGS
options. It works but I'm not convinced that it would keep working after
a Homebrew upgrade. The worst case scenario in this system is "burn down
the Python with `pyenv remove 3.11.7`, re-run `make deps`" so it's not that
inconvenient, but it's not ideal. This is probably a reason why — if not
the reason why — pyenv discourages linking against Homebrew on Linux.