Clarify make usage slightly
I noticed that the `make help` target had some ANSI escape sequences
that didn't seem to do anything:

    @awk 'BEGIN {FS = ":.*##"; printf "\nUsage:\n  make \033[36m\033[0m\n"}

Those escape sequences at the end mean "change the color to blue, then
change it back to normal".

I've seen this pattern copy-pasted around the web for a while, so I was
curious to find what it was originally. I did some historical
spelunking. The earliest example I can find that closely matches this
help target is this:
https://www.thapaliya.com/en/writings/well-documented-makefiles/

In this version, there used to be a `<target>` inside the color
sequences. I can see why that might be helpful. Otherwise, why have the
first line that prints make at all?